Convert xc-fasm2bels to use fpga-interchange library.
