+++ .gitmodules
@@ -4,3 +4,6 @@
 [submodule "third_party/prjxray"]
 	path = third_party/prjxray
 	url = https://github.com/SymbiFlow/prjxray.git
+[submodule "third_party/python-fpga-interchange"]
+	path = third_party/python-fpga-interchange
+	url = https://github.com/SymbiFlow/python-fpga-interchange.git
