+++ fasm2bels/models/verilog_modeling.py
@@ -31,7 +31,7 @@ import re
 import fasm
 from fasm2bels.make_routes import make_routes, ONE_NET, ZERO_NET, prune_antennas
 from fasm2bels.database.connection_db_utils import get_wire_pkey, create_maybe_get_wire
-from fasm2bels.lib.physical_netlist import create_site_routing
+from fasm2bels.lib.interchange import create_site_routing
 
 
 def pin_to_wire_and_idx(pin):
