+++ requirements.txt
@@ -7,5 +7,6 @@ progressbar2
 pyyaml
 -e git://github.com/SymbiFlow/prjxray.git#egg=prjxray
 -e git://github.com/SymbiFlow/fasm.git#egg=fasm
--e git://github.com/SymbiFlow/symbiflow-rr-graph.git#egg=symbiflow-rr-graph
+-e git://github.com/SymbiFlow/symbiflow-rr-graph.git#egg=rr-graph
+-e third_party/python-fpga-interchange/
 yapf==0.24.0
